+++ android/app/src/main/kotlin/com/example/output_route_selector/MainActivity.kt
@@ -1,5 +1,0 @@
-package com.example.output_route_selector
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity : FlutterActivity()